+++ src/Factors/include/KinDynVIO/Factors/CentroidalDynamicsMeasurementBias.h
@@ -26,7 +26,6 @@ namespace gtsam
 class CDMBiasCumulative
 {
 private:
-    gtsam::Vector3 m_biasGyro;
     gtsam::Vector3 m_biasCOMPositionInBase;
     gtsam::Vector3 m_biasNetExternalForceInBase;
     gtsam::Vector3 m_biasNetExternalTorqueInBase; // contains net effect of contact force measurement lever arm biases and contact torque measurement biases
@@ -36,15 +35,13 @@ public:
     inline static const std::size_t dimension = 12;
 
     CDMBiasCumulative();
-    CDMBiasCumulative(const gtsam::Vector3& bGyro,
-                      const gtsam::Vector3& bCOMPositionInBase,
+    CDMBiasCumulative(const gtsam::Vector3& bCOMPositionInBase,
                       const gtsam::Vector3& bNetForceInBase,
                       const gtsam::Vector3& bNetTorqueInBase);
-    CDMBiasCumulative(const gtsam::Vector12& v);
+    CDMBiasCumulative(const gtsam::Vector9& v);
 
-    gtsam::Vector12 vector() const;
+    gtsam::Vector9 vector() const;
 
-    const gtsam::Vector3& gyroscope() const;
     const gtsam::Vector3& comPositionInBase() const;
     const gtsam::Vector3& netExternalForceInBase() const;
     const gtsam::Vector3& netExternalTorqueInBase() const;
@@ -54,18 +51,15 @@ public:
     // H2 = df(y, b) / dy
     // where f(y, b) = y - bi
     // bi is the i-th bias selector in the set of biases
-    // serialized as bg,bcom, bf,btau
-    gtsam::Vector3 correctGyroscope(const gtsam::Vector3& measurement,
-                                    gtsam::OptionalJacobian<3, 12> H1 = boost::none,
-                                    gtsam::OptionalJacobian<3, 3> H2 = boost::none) const;
+    // serialized as bcom, bf,btau
     gtsam::Vector3 correctCOMPositionInBase(const gtsam::Vector3& measurement,
-                                            gtsam::OptionalJacobian<3, 12> H1 = boost::none,
+                                            gtsam::OptionalJacobian<3, 9> H1 = boost::none,
                                             gtsam::OptionalJacobian<3, 3> H2 = boost::none) const;
     gtsam::Vector3 correctNetExternalForceInBase(const gtsam::Vector3& measurement,
-                                                 gtsam::OptionalJacobian<3, 12> H1 = boost::none,
+                                                 gtsam::OptionalJacobian<3, 9> H1 = boost::none,
                                                  gtsam::OptionalJacobian<3, 3> H2 = boost::none) const;
     gtsam::Vector3 correctNetExternalTorqueInBase(const gtsam::Vector3& measurement,
-                                                  gtsam::OptionalJacobian<3, 12> H1 = boost::none,
+                                                  gtsam::OptionalJacobian<3, 9> H1 = boost::none,
                                                   gtsam::OptionalJacobian<3, 3> H2 = boost::none) const;
 
     /** print with optional string */
@@ -84,7 +78,7 @@ public:
     CDMBiasCumulative operator-() const;
 
     /** addition of vector on right */
-    CDMBiasCumulative operator+(const gtsam::Vector12& v) const;
+    CDMBiasCumulative operator+(const gtsam::Vector9& v) const;
 
     /** addition */
     CDMBiasCumulative operator+(const CDMBiasCumulative& other) const;
@@ -100,7 +94,6 @@ private:
     friend class boost::serialization::access;
     template<class ARCHIVE>
     void serialize(ARCHIVE & ar, const unsigned int /*version*/) {
-        ar & BOOST_SERIALIZATION_NVP(m_biasGyro);
         ar & BOOST_SERIALIZATION_NVP(m_biasCOMPositionInBase);
         ar & BOOST_SERIALIZATION_NVP(m_biasNetExternalForceInBase);
         ar & BOOST_SERIALIZATION_NVP(m_biasNetExternalTorqueInBase);